requirements.txt
awesome-slugify==1.6.5
# Update version below to latest from https://pypi.python.org/pypi/hdx-python-api/
hdx-python-api==2.2.8
